requirements_video.txt
python-pptx>=0.6.21
Pillow>=10.0.0
moviepy>=1.0.3
gtts>=2.4.0
pygame>=2.5.0







